refactor: fcm device token generation (#116)
