+++ messagingpush/src/main/java/io/customer/messagingpush/di/DiGraphMessagingPush.kt
@@ -1,7 +1,7 @@
 package io.customer.messagingpush.di
 
-import io.customer.messagingpush.provider.FCMTokenProvider
 import io.customer.messagingpush.provider.FCMTokenProviderImpl
+import io.customer.sdk.device.DeviceTokenProvider
 import io.customer.sdk.di.CustomerIOComponent
 
 /*
@@ -10,5 +10,5 @@ This file contains a series of extensions to the common module's Dependency inje
 The use of extensions was chosen over creating a separate graph class for each module. This simplifies the SDK code as well as automated tests code dramatically.
  */
 
-internal val CustomerIOComponent.fcmTokenProvider: FCMTokenProvider
-    get() = override() ?: FCMTokenProviderImpl(logger)
+internal val CustomerIOComponent.fcmTokenProvider: DeviceTokenProvider
+    get() = override() ?: FCMTokenProviderImpl(logger = logger, context = context)
